requirements.txt
Flask==2.2.3
mysql-connector-python==8.0.32
Werkzeug==2.2.3
python-dotenv==1.0.0
Flask-WTF==1.1.1
email-validator==2.0.0
numpy==1.24.3
pandas==1.5.3
openpyxl==3.1.2
xlsxwriter==3.1.2
matplotlib==3.7.1
requests==2.31.0
beautifulsoup4==4.12.2
